Pimpl most of the settings in GlobalSettingsPanel to reduce header pollution.

+++ SerialPrograms/Source/CommonFramework/ImageTools/WaterfillUtilities.cpp
@@ -6,9 +6,10 @@
 
 #include <map>
 #include "Common/Cpp/Color.h"
-#include "CommonFramework/GlobalSettingsPanel.h"
 #include "Kernels/Waterfill/Kernels_Waterfill_Session.h"
 #include "Kernels/Waterfill/Kernels_Waterfill_Types.h"
+#include "CommonFramework/GlobalSettingsPanel.h"
+#include "CommonFramework/Logging/Logger.h"
 #include "CommonFramework/ImageMatch/WaterfillTemplateMatcher.h"
 #include "CommonFramework/ImageTools/BinaryImage_FilterRgb32.h"
 #include "CommonFramework/ImageTypes/ImageRGB32.h"
@@ -32,8 +33,8 @@ std::pair<PackedBinaryMatrix, size_t> remove_center_pixels(
 //    cout << matrix.dump() << endl;
 
     //  Sort all pixels by distance from center.
-    size_t center_x = object.center_of_gravity_x() - object.min_x;
-    size_t center_y = object.center_of_gravity_y() - object.min_y;
+    size_t center_x = (size_t)(object.center_of_gravity_x() - object.min_x);
+    size_t center_y = (size_t)(object.center_of_gravity_y() - object.min_y);
     std::map<uint64_t, size_t> distances;
     for (size_t r = 0; r < height; r++){
         for (size_t c = 0; c < width; c++){
