+++ SerialPrograms/Source/PokemonSV/Programs/Farming/PokemonSV_MaterialFarmerTools.cpp
@@ -662,7 +662,7 @@ void fly_from_paldea_to_blueberry_entrance(const ProgramInfo& info, ConsoleHandl
         // move cursor to Blueberry academy fast travel point (up-left)
         // try different magnitudes of cursor push with each failure.
         int push_magnitude = 81 + adjustment_table[numAttempts];
-        pbf_move_left_joystick(context, 0, 0, push_magnitude, 50);
+        pbf_move_left_joystick(context, 0, 0, (uint16_t)push_magnitude, 50);
 
         // press A to fly to Blueberry academy
         isFlySuccessful = fly_to_overworld_from_map(info, console, context, true);
@@ -840,7 +840,7 @@ void fly_from_blueberry_to_north_province_3(const ProgramInfo& info, ConsoleHand
         // move cursor up-left
         // try different magnitudes of cursor push with each failure.
         int push_magnitude = 168 + adjustment_table[numAttempts];
-        pbf_move_left_joystick(context, 112, 0, push_magnitude, 50);
+        pbf_move_left_joystick(context, 112, 0, (uint16_t)push_magnitude, 50);
 
         // press A to fly to North province area 3
         isFlySuccessful = fly_to_overworld_from_map(info, console, context, true);
